Split out arithmetic code
