+++ packages/language/src/utils.ts
@@ -1,0 +1,5 @@
+export function assertTrue(condition: boolean): asserts condition {
+    if(!condition) {
+        throw new Error('Condition is supposed to be true!');
+    }
+}
